cast away the size_t to the size we're printing (thx emton)
svn:revision:36544
svn:branch:trunk
svn:account:brlcad

+++ src/libbu/malloc.c
@@ -502,7 +502,7 @@ bu_prmem(const char *str)
 	}
     }
 
-    fprintf(stderr, "%lu allocation entries\n", count);
+    fprintf(stderr, "%lu allocation entries\n", (unsigned long)count);
 
 
 }
